+++ src/main.js
@@ -46,6 +46,7 @@ export { default as DTextarea } from "./d-textarea/DTextarea.vue";
 export { default as DTextfield } from "./d-textfield/DTextfield.vue";
 export { default as DTooltip } from "./d-tooltip/DTooltip.vue";
 export { default as DWysiwyg } from "./d-wysiwyg/DWysiwyg.vue";
+export { default as DCopyInput } from "./d-copy-input/DCopyInput.vue";
 export { default as DarkModeProvider } from "./providers/DarkModeProvider.vue";
 export { default as ToastProvider } from "./providers/ToastProvider.vue";
 export { default as ThemeProvider } from "./providers/ThemeProvider.vue";